requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
idna==3.4
IPy==1.1
maltiverse==1.0.0
oauthlib==3.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
shodan==1.28.0
six==1.16.0
slack-sdk==3.19.5
urllib3==1.26.13
XlsxWriter==3.0.3
